Remove game() and implement button choices
In this commit, I removed the game() function that made the game play a
certain amount of times. The game can now run an unlimited amount of
times and the user input is now done by pressing 3 different buttons for
all possible choices.

Additionally, a scoreboard was implemented on index.html and its
tracking is going to be implemented on the next commit. In addition,
playRound() was refactored to render the output on the webpage as
opposed to print it to the console. This was necessary in order to
display the results on the screen. And the reason why we removed game()
is because we no longer need to run playRound() a certain amount of
times to get some output. Rather, the function now runs every time a
button is pressed.