Fix instant reset on scoreboard bug
Calling updateScore() at the end of resetScores() made sense at first
since all the scores would be set to zero and then automatically update
the scoreboard as well.

While that worked as intended, it caused a bug where the score would
go back to zero and the player could not see the final results. As a
result, I decided to move the call to update the score to the button
event listeners. It makes more sense to reset the scores once the player
clicks on a button rather than right away.

This may change on a future commit once I add more style to the page. I
am thinking about creating a modal to display the results and ask the
player if they want to play again.