Improve scoreboard by dividing player from COM
In this commit, several improvements are made to the application.
First, the UI's background colors are reworked and rounded borders
are added to the main content and around the scoreboard.

Previously, the scoreboard was represented as two separate values
contained within the same element. This was refactored so that the
player and computer scores are now separate and rendered individually.
They are still updated at the same time by updateScores(), but their
concerns are now separate. This makes it easier to both align the UI
elements and to change their style at a later time if needed.

Finally, updateScore() was refactored into updateScores(). This function
now updates the value of the newly added player and COM scores as
opposed to just updating a single value.

The next commit will deal with reworking the default style for the rock,
paper, and scissor buttons.