Add addBlur(), removeBlur() and center modal
In this commit, I made several styling changes to the modal. I tried
centering the modal through Flexbox but was unable to. The only solution
I could think of was to use absolute positioning. I found online that in
order to do this, I had to set the top and left margins to 50% and then
use a transform to center it as just setting the margins positioned the
modal slightly out of place.

I also had to resort to implement two methods to add and remove blur
from a target item since changing the blur property of the :modal
pseudoclass did not achieve what I wanted. I wanted the body to blur out
while the modal was open. So, what I did to accomplish this was to add a
CSS class that would blur out an element. These two methods are called
on askForReplay() and on the listeners for the modal (dialogPrompt).

The next commit will deal with styling the contents inside the modal.