Implement sounds on buttons and on dialog
In this commit, I wrote the behavior for all the audio files. Audio was
given to the player buttons when they are hovered over. Sounds for the rock,
paper, and scissors buttons were added. These sounds are played when the
player clicks on them. Finally, sounds were given to the win-lose
scenarios. These sounds are played when the dialog modal pops up at the
end of the 5 games.

One consideration I had to keep in mind was that there is some
repetitive code when playin the fies. One would think that writing a
simple function to encapsulate this behavior would work fine. However, I
found out that doing this would not play the files consistently.
Sometimes they would play, sometimes they would not.

I tried to check for null elements, but this did not work out. I also
tried to check if the file was playing or if it had a time duration
other than zero, but this did not work out either. As such, I had to
leave the repetitive code that plays the audio files in their respective
place.

In the next commit, I intend to make some adjustments to the UI so that
the footer stays in place and that the player doesn't have to scroll
down if they so wish to see it. I also personally feel that doing so
gives the front page a better appearance. In addition, I also intend to
refactor playRound() to decouple the round logic from the rendering
logic. Currently, the function renders the results to the result
element, and this is a side effect. I want to decouple it and write a
function that specifically renders the output of playRound().