Add renderOnResult() to decouple element rendering
In this commit, I decided to decouple some of the functionality of some
functions. Namely, playRound() and announceWinner() made a direct call
to the result element to render the result.

In order to make these functions more maintainable, I removed that
functionality and ported it over to renderOnResult(). This still
produces the same result and makes it easier to debug the program in
case it is ever needed.

Some functions could still be reworked to reduce side effects, but I
feel that doing so may make the code harder to understand.

On future commits, I would like to completely rework the UI to make it
more lively by adding more animations to the game.